requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.9.0
pydantic-settings>=2.5.0
supabase
python-jose[cryptography]
passlib[bcrypt]
loguru
pdfplumber
python-multipart
httpx
python-dotenv
